Refatorando service and utils
