+++ src/connection/DAOFactory.java
@@ -1,4 +1,4 @@
-package utils;
+package connection;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
