+++ src/controllers/IServiceCliente.java
@@ -1,13 +1,13 @@
-package utils;
+package controllerslers;
 
 import java.sql.SQLException;
 import java.util.List;
 
 import model.Cliente;
 
-public interface IDAOCliente {
-    public abstract void adicionaCliente(Cliente c) throws SQLException;
+public interface IServiceCliente {
+    public abstract void cadastraRegistro(Cliente cliente) throws SQLException;
     public abstract List<Cliente> getLista(String nome) throws SQLException;
-    public abstract void altera(Cliente c) throws SQLException;
     public abstract void remove(int id) throws SQLException;
+    public abstract void altera(Cliente cliente) throws SQLException;
 }
