+++ src/controllers/IServiceLivro.java
@@ -1,11 +1,11 @@
-package utils;
+package controllers;
 
 import java.sql.SQLException;
 import java.util.List;
 
 import model.Livro;
 
-public interface IDAOLivro {
+public interface IServiceLivro {
     public abstract void adicionaLivro(Livro livro) throws SQLException;
     public abstract List<Livro> getLista(String exemplar) throws SQLException;
     public abstract void altera(Livro livro) throws SQLException;
