+++ src/controllers/IServiceMulta.java
@@ -1,11 +1,11 @@
-package service;
+package controllers;
 
 import java.sql.SQLException;
 import java.util.List;
 
 import model.Multa;
 
-public interface IServiceD {
+public interface IServiceMulta {
     public abstract void adicionaMulta(Multa multa) throws SQLException;
     public abstract List<Multa> getLista(String id) throws SQLException;
     public abstract List<Multa> getListaMultaPorCliente(String id_cliente) throws SQLException;
