+++ src/dao/DAOEmprestimo.java
@@ -1,4 +1,4 @@
-package utils;
+package dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -6,6 +6,8 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+
+import connection.DAOFactory;
 import model.Emprestimo;
 
 public class DAOEmprestimo implements IDAOEmprestimo {
