+++ src/dao/IDAOMulta.java
@@ -1,4 +1,4 @@
-package utils;
+package dao;
 
 import java.sql.SQLException;
 import java.util.List;
